requirements.txt
opencv-python
pillow
python-pptx
pytesseract 